frontend test fixes
