+++ app/webpack/reactjs/components/open_studios/special_event_schedule_fields.tsx
@@ -1,6 +1,5 @@
 import { DateTime } from "luxon";
 import { MauCheckboxField } from "@reactjs/components/mau_checkbox_field";
-import { parseTimeSlot } from "@js/app/time_utils";
 import * as types from "@reactjs/types";
 import React, { FC } from "react";
 
@@ -18,10 +17,17 @@ interface TimeSlotCheckBoxProps {
   name: string;
 }
 
+export const parseTimeSlot = (timeSlot: string): types.TimeSlot => {
+  const [start, end] = timeSlot
+    .split("::")
+    .map((v) => DateTime.fromSeconds(parseInt(v, 10)));
+  return { start, end };
+};
+
 const generateSlotFieldName = (timeslot: string): string =>
   `videoConferenceSchedule[${timeslot}]`;
 
-const formatTimeSlot = (timeslot: types.TimeSlot): string => {
+export const formatTimeSlot = (timeslot: types.TimeSlot): string => {
   const { start, end } = timeslot;
   return `${start.toLocaleString(DateTime.TIME_SIMPLE)} - ${end.toLocaleString(
     DateTime.TIME_SIMPLE
@@ -30,7 +36,13 @@ const formatTimeSlot = (timeslot: types.TimeSlot): string => {
 
 const TimeSlotCheckBox: FC<TimeSlotCheckBoxProps> = ({ timeslot, name }) => {
   const label = formatTimeSlot(timeslot);
-  return <MauCheckboxField classes="open-studios-info-form__special-event-schedule__timeslot" label={label} name={name} />;
+  return (
+    <MauCheckboxField
+      classes="open-studios-info-form__special-event-schedule__timeslot"
+      label={label}
+      name={name}
+    />
+  );
 };
 
 export const SpecialEventScheduleFields: FC<SpecialEventScheduleFieldsProps> = ({
@@ -42,7 +54,10 @@ export const SpecialEventScheduleFields: FC<SpecialEventScheduleFieldsProps> = (
   const slots = specialEvent.timeSlots;
 
   return (
-    <div className="open-studios-info-form__special-event-schedule">
+    <div
+      className="open-studios-info-form__special-event-schedule"
+      data-testid="open-studios-info-form__special-event-schedule"
+    >
       <div className="open-studios-info-form__special-event-schedule__label">
         I will be open for virtual visitors the following hours.
       </div>
